Commit final

+++ src/MediaAlunos.java
@@ -9,12 +9,12 @@ public class MediaAlunos {
         public static void main(String[] args) {
 
             System.out.println("Informe as quatro notas dos alunos:");
-            int nota1 = 5;
-            int nota2 = 6;
-            int nota3 = 7;
-            int nota4 = 8;
+            double nota1 = 5;
+            double nota2 = 6;
+            double nota3 = 7;
+            double nota4 = 8;
 
-            int media = (nota1 + nota2 + nota3 + nota4) / 4;
+            double media = (nota1 + nota2 + nota3 + nota4) / 4;
 
             System.out.println("A média dos alunos é: " + media);
